+++ test/conformance/suite_test.go
@@ -174,6 +174,7 @@ func startControllerManager(metadata metadata.Info) <-chan struct{} {
 	cfg.ValidatingWebhookEnabled = false
 	cfg.AnonymousReports = false
 	cfg.ClusterCAKeyType = mgrconfig.ECDSA
+	cfg.GatewayAPIExperimentalEnabled = true
 
 	cfg.NewClientFunc = func(config *rest.Config, options client.Options) (client.Client, error) {
 		// always hijack and impersonate the system service account here so that the manager